+++ packages/service-fastify-cookie/src/hooks.js
@@ -1,0 +1,5 @@
+const SERVICE_NAME = `fastify-cookie`;
+
+module.exports = {
+  SERVICE_NAME,
+};