+++ packages/kitchensink/CHANGELOG.md
@@ -7,4 +7,7 @@
 - Added examples for `@forrestjs/service-jwt`
 - Added examples for `@forrestjs/service-hash`
 - Added examples for `@forrestjs/service-fetchq`
+- Added examples for `@forrestjs/service-fetchq-static`
+- Added examples for `@forrestjs/service-fetchq-cors`
+- Added examples for `@forrestjs/service-fetchq-cookie`
 - Added `app01` examples for a complex app